Refactor to hooks
